refactor detailsPage on store
