+++ src/client/src/containers/PageAddServices/ServiceCard/index.tsx
@@ -14,8 +14,8 @@ import { isLoggedInSelector } from "../../../store/config/azure/selector";
 import { AppState } from "../../../store/combineReducers";
 import { ReactComponent as PriceSVG } from "../../../assets/money.svg";
 import { ReactComponent as TimeSVG } from "../../../assets/timer.svg";
-import { setDetailPageAction } from "../../../store/navigation/routes/action";
 import { ROUTE } from "../../../utils/constants/routes";
+import { setDetailPageAction } from "../../../store/config/detailsPage/action";
 
 interface IProps {
   service: IService;
