+++ src/client/src/store/config/detailsPage/routesActionType.ts
