+++ src/client/src/store/config/detailsPage/visitedAndSetPageReducer.ts
@@ -1,6 +1,6 @@
 import { ROUTE } from "../../../utils/constants/routes";
 import { IRoutes } from "../../userSelection/pages/model";
-import { NAVIGATION_ROUTES_TYPEKEYS } from "../typeKeys";
+import { NAVIGATION_ROUTES_TYPEKEYS } from "../../navigation/typeKeys";
 import RootAction from "../../ActionType";
 
 const initialState = {
