+++ src/client/src/store/navigation/combineReducers.ts
@@ -1,11 +1,9 @@
 import { combineReducers } from "redux";
 import modals from "./modals/combineReducers";
-import routes from "./routes/combineReducers";
 import routesNavItems from "./routesNavItems/reducer";
 
 const navigationReducer = combineReducers({
   modals,
-  routes,
   routesNavItems
 });
