--- src/client/src/store/navigation/routes/combineReducers.ts
@@ -1,6 +1,0 @@
-import { combineReducers } from "redux";
-import {isVisited, selected} from "./visitedAndSetPageReducer";
-
-export default combineReducers({
-  isVisited, selected
-});
